langchain-olamma/pyproject.toml
[tool.poetry]
name = "langchain-ollama-qwen"
version = "0.1.0"
description = "Example Python app to call Ollama qwen2.5:0.5b model"
authors = ["Example <example@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ollama = "^0.1.0"
requests = "^2.28"
python-dotenv = "^1.0"
typer = "^0.9"
rich = "^13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
